backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
beautifulsoup4
pymupdf
openai
python-dotenv
playwright
